Rename class GetValidationUsecaseAll to GetUsecaseCollectionForFile for code clarity
