+++ usecase/src/main/java/org/mobilitydata/gtfsvalidator/usecase/GetUsecaseCollectionForFile.java
@@ -18,6 +18,8 @@
 
 package org.mobilitydata.gtfsvalidator.usecase;
 
+import org.mobilitydata.gtfsvalidator.domain.entity.FileSpecificUsecase;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -25,10 +27,10 @@ import java.util.Map;
 /**
  * Use case to retrieve the list of use cases to execute in order to process to validation of GTFS dataset
  */
-public class GetValidationUsecaseAll {
-    private final Map<String, List<ValidationUsecase>> usecaseByFilename;
+public class GetUsecaseCollectionForFile {
+    private final Map<String, List<FileSpecificUsecase>> usecaseByFilename;
 
-    public GetValidationUsecaseAll(final Map<String, List<ValidationUsecase>> usecaseByFilename) {
+    public GetUsecaseCollectionForFile(final Map<String, List<FileSpecificUsecase>> usecaseByFilename) {
         this.usecaseByFilename = usecaseByFilename;
     }
 
@@ -39,8 +41,8 @@ public class GetValidationUsecaseAll {
      * @param filenameListToProcess the list of files to validate
      * @return the list of use cases to execute
      */
-    public List<ValidationUsecase> execute(final List<String> filenameListToProcess) {
-        final List<ValidationUsecase> toReturn = new ArrayList<>();
+    public List<FileSpecificUsecase> execute(final List<String> filenameListToProcess) {
+        final List<FileSpecificUsecase> toReturn = new ArrayList<>();
         filenameListToProcess.forEach(filename ->
                 toReturn.addAll(usecaseByFilename.get(filename)));
         return toReturn;
